pretty big commit here, only because I restricted the rendering engine inside systemiv/lib so there is a lot of files changed but these are just new include paths
refactored the rendering engine so that the 2d rendering and the 3d rendering is in separate folders
I think it was needed after the 3d globe view was making the old render folder huge

implemented batching to the 3d globe, I just duplicated the 2d buffers file and then modified the implementations slightly to work with the 3d globe

fixed the elevation of the units, fighters and bombers now fly above the globe by a good bit, the unit trails also match the elevation
the 2d unit trails have also been resurrected as the solid line trails was extremely distracting

nuke trails now are segmented lines with a good bit of space between them, updated movingobject.cpp to give us a much better sample rate so the trajectory lines are not jagged
this needed all of the old trail logic to be modified as we now have more history to parse with the increase segment sizes

the playback controls window is now a fading window so that its not an eye sore to people, it matches the logic of the chat window and the toolbar

I also corrected the alpha blending issue with the 3d globe sprites, now they match the 2d transparency. before all sprites were solid and had z fighting issues

removed sprite nuke rendering in the 3d globe mode, now we use a 3d model which is not perfect but its better than the billboarded nuke sprite
still having issues with the nuke angle following the actual trajectory. the calculation that is done for 2d does not work here

the debug menu also works now in 3d mode, this was as simple as duplicating the logic that the 2d menu uses to track draw calls
added a little indicator to the menu title to show what draw calls we are tracking

changed the memory initialisation of the replay viewer inside the cmakelists to help dogshit pcs not error out