updated globe controls and the caching system for user downloads adjusted the camera rotation speeds for the mouse and with wasd after some people reported it felt too slow for wasd and too fast for middle mouse, personally I think its too slow but I am using 400 dpi most people are sane humans and have 1600 dpi
the playback control window has been refactored to have the buttons on the left side of the window,
alongside this change I added two new buttons. one that toggles the ui on or off, or the globe on
or off.

updated the server.js and cmakelists to compress the .data folder and decompress when the user loads
the replay viewer, it 100 percent feels faster than before. but we still have strange issues with the
loading of the replay viewer for some users

a worrying discovery led me into a rabbit hole of sync log tracking, the result is that the steam client
and the community source even after the sync compatibility patch is still not 100 percent sync compatible
in certain edge cases. specifically when there is ai involved. I observed a 3v3 which had game state wander
when victory timer started but this is the only real multiplayer game to have this issue. after some close
inspection it actually looks like the syncrand() calls sometimes get flipped between units which points to
compiler optimisations possibly destroying some loops and functions. or more likely there is a slight edge
case that we still have to fix