Added FPS limit and UI scaling added an fps limit inside the screen options menu, this is good as now we are running such high fps that we dont want our gpu to be overworking all the time. so now we can set it to whatever we want
implemented a cool way to implement UI scaling, we actually just create two different resolutions
we store our real resolution seperate from the games internal resolution, when we change our ui scale
in the menu this actually changes defcons internal resolution. this tricks defcon into rendering everything
at different sizes, of course the mouse coordinate system had to be modified to work with this

fixed the misalignment with nukesymbol.bmp on land units when nukes are targetting them, this was due to my
ignorance when i fixed this issue for cities. now the fix only applies to cities and not land units

added more memory to the emscripten release build as with all the changes we have made defcon is more memory
intensive so we had to increase the memory allocation from 512mb to 1024mb

in progrss but setting defcon as dpi aware causes issues when changing resolutions in fullscreen to non native
it requires a restart or else the game will render as a box in the bottom right of the screen. ah well