changes to line widths, added smooth lines and endgame configuration since we know are using compatiblity opengl profile we can now use smooth lines again for defcon this was added back to the graphics options in the menu. of course this is only for desktop
i estimate the line widths of the unit trails in steam defcon is 1.5f so i added this to the
flush functions inside the buffers section of the renderer to avoid immediate mode rendering
from overriding the state of the linewidth

added endgame configuration, what this does is whenever the expanded client is released to the
public it will make real multiplayer games non sync compatible. but allows the community source
to be sync compatible. this means that when we watch recordings the parameters match steam and
when we are not watching a recording it will revert back to community parameters