massive globe rendering overhaul, this is a big one lets first start of with some low level stuff, i changed the fog program to preserve alpha channels for orientation based fog, now the globe doesnt look too bright and now gunfire and trails preserve their alpha channel when fog is applied
alongside this i increased all the 3d buffers to a more reasonable size to prevent constant
flushing when its not needed

i changed the include path for emscripten in universal include for linking against opengl
as i made a mistake which made emscripten include glew.

shrunk down the size of the data and mods folder for the emscripten build to reduce
download times for slower internet people. every single MB counts here

we still have an ongoing bug with the replay functionality, sometimes if we leave the
recording server and try to watch another recording the world or game object never
gets deleted. its a rare bug but im keeping an eye on it

i also refactored the scene rendering in render3dglobe(), we changed some thing about
specifically we added seperate scenes so we can seperate fog rendering
i also tweaked some of the fog values for a better look

i changed the max zoomed out level to 3 instead of 10 as nobody ever will use 10
this means the globe starting zoom level is now the max you can zoom out

we also moved the gridlines rendering to inside render3dglobe instead of render3dscene
the gridlines now have more segments built than before, so now gridlines lay flatter
and smoother on the globe than before.

i finally added real culling to the world objects on the globe, we just dont write
color information when the objects are behind the globe surface
a huge improvement to frustrum culling which was ass

now onto the good stuff
the globe now supports

unit ghost rendering
launch detections
nuke target rendering
orders overlay rendering

the ghost rendering was important to nail as we need the functionality
to be identical to 2d mode, if its not then people can cheat using globe
mode in real games. for this reason im not calling it a success yet but
so far the functionality looks identical

launch detections are almost perfect, but there is still some strange behaviour
that i need to rule out, specifically the nuke symbols do not dissapear after
their timer from time to time during testing

the nuke targets work the same as the launch detection rendering, they use the
same bmp and their logic is just flipped so that we can see the nuke symbols
on our targets.

the orders overlay was a really neat feature, we use great circle calculations
for the actual order lines as we had the long lat flipping issue without this.
we have special handling for nukes, we only render the cursor target bmp for
the nukes in flight since the order lines follow a straigt path along the globe
and not in the sky

some other tweaks that was made is i have now moved the unit elevation calculations
to a seperate function, but for now we have all the units the same Z level off the
globe as i really didnt like the look of different elevations

alongside this i also made the action lines, unit trails and gunfire slightly below
all the units to prevent alpha Z fighting issues

the 3d nuke rendering and the nuke trails now reside inside their own functions
for better code organisation, i really do not like messy code so this felt like
a necessity