fixed sync compatiblity error in rare edge cases the emscripten replay viewer and the 64 bit windows target had a rare edge case that caused sync incompatiblity with 1 recording after some digging i found that the floating point math optimisations were too aggressive for emscripten and x64 windows so removing these optimisations fixes the sync errors completely
dont be scared about the emscripten files, this is being added as it should mean once you clone the repo emscripten does not
need to be setup again. it should just work when you click the wasm build batch files

i also fixed a small issue that prevented the emscripten target from building in window manager sdl, it was defining glew for
emscripten when linux was defined. that is now fixed

the version of the replay viewer is now 1.97.4