complete map renderer overhaul, removed debug menu and fixed scene managment first fix was actually removing the debug menu and adding it to app.cpp
then i started to replace all the immediate mode calls to use our new UI functions, this worked okay for most things
but some functions like the renderradar function gave me trouble with depth testing. so it still uses RectFill for the
world darkening filter

speaking of renderradar, i moved it outside the batching context in maprenderer::render()
since the radar function has its own micro batching for the circles to join together we
dont need this anymore
reason for this change is because the world darken filter was actually only working for cities
and not every other unit. that is now fixed

i moved the renderworldmessages function inside the first batched context, wrapped in a begin and end
call for the text buffer

so all world messages are 1 draw calls

map renderer is now at an acceptable level for my standards now
with UI enabled and a 3v3 playing with all the overlays on we get 100 draw calls
and no overlays around 60 draw calls