added some checks to prevent certain ui elements from rendering in replay mode first change was to make the world status bar at the top of the map renderer scene behave the same way as if we are a spectator, so users dont try to play with the buttons
then for the alliances window we add the same checks here. as in radar perspective mode you can actually see the
windows appear and if you click back to spectator when these windows are shown the game crashes because it tries
to access a teamid that doesnt exist

the same changes were made to the votingsystem windows so these will never appear either

the sidepanel no longer renders when in replay mode, before you could access it in radar mode and actually attempt
to place units. this is a bad look on me and its just not professional

the toolbar was updated to reflect these changes. now the units and alliance buttons do not render in replay mode
so that users cannot even attempt to access these windows

and lastly i replaced all the legacy calls in these files to use the new batched functions