new custom client for silo sync practicing, currently in testing phase this custom client is able to calculate flight times from each silo and assign rankings to them when the user launches their silos a timer will start on the next silo to tell the user when to launch next.
the system has some defensive programming to prevent some edge cases and users fucking around but im
sure people will still find ways to break it :)

this client is similiar to the replay viewer in that it prevent multiplayer and loads us directly into a
local game lobby. it prevents switching to spectator and gamemode. however the gamemode options are
available to be tinkered with

max teams is set to 2 for this client so the user can practice their syncs against 1 ai player only
we dont want to give the user too much gameplay as in the future an emscripten version will come out

the user and AI player is bound to just placing radars and silos to further drive the point that this
client is purely for testing your syncs out not for playing the game

to make all of this work i added a preprocessor SILO_PRACTICE which isolates the main client from these
features.

one really imporant thing i done to prevent abuse or cheating i make generatesyncvalue return 0 in this
custom client. instant desyncs with other players
i also prevented the server from advertising on the metaserver, after all these changest even if somebody
types in the ip manually to connect they will be met with an instant desync :)

while i was in there i also fixed up the preprocessors for the RECORDING_PARSING option which was slightly broken
and wouldnt actually compile without it defined