fixed the biggest performance bottleneck in the renderer before i get onto the big changes here is some of the smaller changes
on emscripten we now do not buffer orphan anymore, we just use glsubbufferdata directly as during testing
it actually reduces stuttering for webgl2
on desktop we still use orphaning inside uploadvertexdatatovbo()

now onto the best change to date

so i always knew the VBO system was unoptimised and very old, it dates back to june 11th
the first thing i noticed is we were not using the cached uniforms for the vbos
and we also binded the vbos twice in the constructor. causing unnecessary memory thrashing

instead of using line segments now for building the vbos, this actually duplicated the vertex
array which resulted in double allocation!

we now use indexed drawing with gl line strip which basically halves the memory usage for storing
and building

for context with the old system
with the 3d and 2d vbo combined
we allocated 320mb of ram when loading the high resolution mod
this is now down to 160mb

on windows, we see the most benefit. our fps is higher now on windows compared to emscripten for the first
time ever

the last and most significant change is the dynamic buffer size for the vbos
inside earthdata.cpp we have a function that counts the indicies for the borders
and coastlines and then is able to determine an accurate size for the buffer size

this results in under 200mb of total ram usage for the defcon executable when we do not
have the high resolution mod enabled
compared to 600mb before this change

the last but important change is inside the mod system we now call updategeographyaffectingmods()
to be called during commit() this prevents a bug where if we start the game with a vbo affecting mod
if we disable that mod the vbos will not get rebuilt since the array was empty