added OpenMP threading to resource system in SystemIV all of the sequential resource loading operations are now using OpenMP, i took special care to make sure the threading is safe and does not cause race conditions or data corruption
the result is that mods, and game startup is ALOT faster. the benefit is night and day
loading mods on my machine is practically instant, and game startups is around 1 - 1.5 seconds
faster than single threaded

i guarded all of this functionality with a preprocessor, so the engine does not need openmp to
run.

the biggest win was inside bitmap.cpp and inside sprite_atlas.cpp