Fix use after free bug in RestartSoundLibrary() We had a bug that when we pressed apply in the sound options menu, it would clear the cache of SampleBank which contained pointers to sound files, frequencies etc.
The issue is that we wouldnt actually wait for the sounds currently playing to stop,
which resulted in the game trying to access a sound that no longer existed.

Solution:

We use mutexes to lock the sound instances and allow them to finish playing. Once that
is done, we unlock and free the cache.