Fixed a huge CPU bottleneck in world.cpp with SonarPings While exploring the profiler menu in defcon, i noticed that UpdateRadar() was taking up ALOT of CPU the longer the game progresses.
We would start at around 30 - 50 percent thread usage
then eventually it would rise to obscene levels. Around 60 to 80 percent of the thread time
was being consumed by SonarPings!

The issue was that the array itself was never compacted for m_animations, so over time the array
would grow massive as we would iterate over thousands and thousands of null indexes.
After the fix, UpdateRadar() stays in a reasonable level of thread usage, around 30 - 40 percent.

This fix was extremely simple and i battle tested it to be certain that it caused no sync errors.