Fixed crackle when moving camera, resampled nukerumble and lowered nuke sample group volume Fixed division by zero risk inside the sound system, when moving the camera left and right.
The issue is that when the camera moves fact or even moderately fast, the listener position would approach 0.
When this happens we get ugly crackles and artifacts.

The biggest problem with defcons sound is that the sound files themselves actually contain crackles!
So no matter what we do there will be crackles, lowering the volume inside sounds.txt makes them pretty
much dissapear.

Callum resampled nukerumble as it was really bad. he done a great job