Make RequestFleet synchronous to maintain network determinism... Here is the problem, the old fleet placement logic only calls RequestFleet() 2 times. With the unit placement update we now call RequestFleet() inside MouseUp, this means that we are now calling it 3 times
The result is we bump the fleetId index to start at ID 1 instead of ID 0.
So the solution is now we add a de-duplication check when processing network broadcasts
in the ClientToServer loop.

Alongside this change RequestFleet has been updated to create
a local fleet instantly before we actually build the network packet and send it to other
clients.

This is a hacky workaround that attempts to work with Defcons determinism, not against it
Please do not flame me for this, this is a clean and simple to follow solution that prevents
a massive refactor of the fleet placement logic.