Add Crashpad system to Defcon for Windows We create a .dmp file that can be analysed with whatever tools we need.
We can also send these crash reports to any API endpoint we want, we
also include debug.txt and blackbox.txt if they exist in the API request.
currently its sent to the Expanded website, which the code will be in the
next commit.

Linux support will be next up, however MacOS support will be hard to test without
a Mac...