Removed old camera control system, now it feels more natural and like map renderer... Removed some functions that were poorly written and will need reintegration later, specifically the globe culling system.
Which was responsible for preventing the starfield from being visible through the globe and preventing world objects
from being visible from the front facing hemisphere.

The globe is no longer called from map renderer, its now its own scene and runs independantly from map renderer, the
only system that calls globe is lobby renderer. Which is okay but a refactor to the render loop in defcon could fix that.