Starfield now using Mega-VBO system, added simple culling to world objects... using dot product, its extremely crude and needs some fine tuning but it works well and it actually improves FPS when we are looking at a side of the globe with no world objects.
The last step is figuring out a correct method to cull the nuke trajectories 0__0

Starfield rendering was extreme before, the default star count was 1200, even though we batched the draw calls
into a single one we still had to call StaticSprite3D 1200 times!

It only got worse with more stars, with 20k stars (the max) our cpu was begging for help and our fps was in the
low hundreds in the menu.

Now we generate the starfield, add the textured quads to the VBO and then call it once and render it once. The
starfield now consumes 2 percent of the main thread in the menu compared to 40 percent before these changes. Im
really happy with the result.