MegaVBO system now supports triangle building and storage, i also... refactored the VBO system for 3D. We now have 4 seperate files for each primitive that the VBO builds and stores. Alot better than cramming it all into one file.
The total memory breakdown in the renderer debug menu is now alot more accurate, we now track all the
VBOs that exist in the codebase. Not just the earthdata VBOs.

I think soon i will create a unified VBO class that combined the 2d and 3d systems together, and making
it a seperate class that inherits from Renderer. To me that seems like a better system