Lots of improvements, added a VBO driven culling sphere... After implementing the triangle VBO system i mirrored the way that the starfield VBO is created and used. Since the culling sphere is static geometry that never changes, its the perfect fit.
The culling sphere serves a purpose, the dotproduct culling can only do so much. Its only really
useful for objects on the globe surface, since its culling radius is the same size as the globe.

The culling sphere disables depth testing and color writing, so anything that renders behind the
sphere is completely blocked out. This includes the starfield, 3d nukes and 3d nuke trajectories.

In globe mode we also cull the coastlines and borders to prevent them being slightly visible through
the orientation fog.

Other improvements are as follows

- Removed globe size option in the menu, why would we want to change the globe radius?
- Improved the VBO rebuilding during mod loading
- Made the globe radius a define in the header, alongside the culling radius
- Added back fog control to the gamecamera
- Updated renderer debug menu to show all the VBOs
- Changed the 2D VBO cachekey names to something more canon
- Improved OOP inside the GlobeRenderer class
- Default star count is now 12500, and default star size is 1.5

We are getting dangerously close to the final step, refactoring the unit rendering system,
and adding proper mouse handling. These two will take the longest and will take special care
and thought to get right.