Fix mouse cursor in globe freezing when hiding the UI, catching this issue was hard... I knew it was an OpenGL state issue however i had no idea where to start, using nvidias frame debugger was essential here. Capturing 2 frames, one frame with a working mouse in globe when world objects were not rendered, and 1 frame where the mouse did not work and there was world objects rendering.
The issue is that the 2D renderer has cached array buffers that our 3D renderer was not using for the VBOs, this meant
that the buffers were mismatching between the 3D and 2D renderer.

This was evident when inside the frame debugger where the mouse would bind to buffer 23, however the 2D renderer actually
has the mouse buffer cached in buffer 10. Since its still bound and valid the mouse would freeze and not move.

The hilariously easy fix was just to make the 3D VBOs upload their vertex data to the cached vertex array in the 2D renderer.
Now the 3D and 2D renderer share the same state!

This fixed another issue that caused the UI to flicker when the mouse was visible in globe, it was a simple buffer mismatch.