ModSystem rework, recursive folder searching for mod.txt and smarter memory managment... A small but really helpful addition is adding a 3 folder recursive folder search for mods, we will search 3 subdirectories inside each mod folder to check if mod.txt exists. If not we will continue and assume the mod is broken.
This fixes the notorious issue with people extracting mods into sub directories and not understanding that
defcon expects mod.txt to be 0 folders deep and no sub dirs.

VBO's were always rebuilt when activating a mod, even with the check i added months ago that doesnt invalidate
the VBO's if the mod does not affect them. The issue is that the resource system is restarted anyway so they
are invalidated regardless!

Now we add the VBO's keys to a list of *protected* VBO's which is then fed into the invalidation functions. That
means when the resource system restarts the VBO's are not invalidated as they are in the protected list.

This practically completes the mod system, nothing else needs to be done.