Complete rendering architecture overhaul, rendering state is now correctly shared between 2d and 3d... We now have a central rendering class inside /lib/render/renderer.cpp. This aims to handle OpenGL state between the 2 renderers, as before renderer3d had its own internal state managment that was not always passed into the main renderer.
The 2D and 3D MegaVBO systems are now their own classes, to me it makes sense as these systems have grown pretty large
and pretty sophisticated. We now access these VBO systems with g_renderer2dvbo and g_renderer3dvbo. Using singletons
is generally non canon for modern codebases but it prevents huge refactoring of App code so who cares.

Most of the changes in this commit are not that interesting to be honest, since we have a g_renderer2d and g_renderer3d
singleton now, most of the app code needed updated as it was referencing g_renderer to access renderer2d class methods.

I know that the refactor was a success as the game runs exactly the same as before, really its just alot easier to
maintain now, these changes were for bug prevention and an easier API to work with.