Integrate the MegaVBO system into the model loader, add instanced rendering with shaders... 3D model loading had a major flaw initially, each model that was loaded in Defcon needed matrix tranforms and mesh computation on the CPU each time it was rendered, this caused an obscene level of CPU overhead when we load any more than 10 models in the scene.
Now the system matches big engines such as unreal, godot and unity, mesh is built and cached once and rendered many times with
no CPU overhead. The use of instancing basically matches the batching architecture used by the sprite and primitive system, since
the mesh and geometry is similiar and never changes we can batch them together into groups of draw calls. Instead of rendering each
model one at a time.

The MegaVBO system is really taking shape here, at first it was just for coastlines and borders. Now we use it for pretty much every
static mesh and geometry in Defcon.