Complete the instancing API, which now shares the same design pattern as... the batching system for sprites and primitives. Its not perfect but its extremely easy to use at the app level compared to before, which is really all that matters.
The instancing system is part of the MegaVBO3D class which makes sense as it closely works with the trianglevbo
module.
The architecture is as follows

Resource -> Model3D -> MegaVBO3D -> Instancing -> Renderer3D -> App

The vertex shader that is used for instancing only supports 64 matrices and colors so that limits our total
instances to that number, we can increase the number in the shader but all gpus support 64 so its best to
maintain compatibility with all machines as best we can.

Smaller changes are

- Disable instancing for the non-instanced ModelShaderUniforms function
- Better usage and initialisation for MAX_INSTANCES
- Removed broken instancing code that existed before