3D Nuke/Trajectories implemented, nukes are perfect but trajectories need some work... 3D nukes are properly instanced which brings the CPU overhead to basically 0, however the trajectories need alot of work which is most likely a result of the renderhistory API.
With lots of nukes in the air the FPS does decrease by a good margin but its not terrible, i get around 400 fps with 130+
nukes in the air, and hte CPU usage in the profiler is over 50 percent just for nuke trajectories. Its something to work on
at some point but really its not the end of the world.

Smaller changes in this update include...

- 2D trajectories are used when the user is not a spectator or in replay mode
- Changed the nuke model scale to something a bit smaller
- Removed duplicate checks for globe mode in the render functions, we return early in globerenderer::render() now
- Added alot more helper functions at the top of the file to reduce code complexity and make it more readable