Removed modsystem dependancy from SystemIV, fixed small edgecase with style mods... >> So mods with custom styles would invalidate vbos and rebuild them incase the colors of these vbos have changed, this is correct and intentionally.
However unlike mods that affect coastlines and international data, when deactivating mods with style changes the vbos
never rebuilt so if the mod changed the colors of the vbos they would stick. So now we do the same checks that geo-
graphy affecting mods do, track the active style mods so when we deactivate them we know to rebuild the vbos.