SystemIV no longer governs filesystem directory structures, the app does... Added more interface code that allows the app to talk to SystemIV, this time its for directory structures. The app must now declare its data root (folder name that SystemIV will access for resources) and also decl- are its subdirectories for the sprite atlas.
The sprite atlas got a well needed makeover, it now builds atlases dynamically based on the directories set
by the app, so now you can have unlimited directories and sprites and spriteatlas will just create an atlas
and coordinates for you for each directory you set.

The system is not finished but its for sure alot better than before, clear seperation between app and engine
code. Which is something that SystemIV heavily lacked before this change.