Refactor the 3D Model system, now supports textures and normals... Felt a bit wrong to just have simple mesh loading in the model system, considering SystemIV is growing to become a pretty solid engine. So i just went back in time to when the model system did support textures and normals and reimplemented it.
Its not finished yet as there are still some architectural inconsitiencies in the model system, such as only
using textures currently and not using normals (normals is supported just not used)

The class hierchy has been completely redone, we now have a Model, ModelBuilder, ModelLoader and ModelMaterial
class. Instead of everything being part of the old Model3D class.

Some smaller changes that have been made are below:
- Made some Vector2 methods available at compile time that werent before
- Updated vertex layout for the Textured3D struct

This was a bit of a side project and doesnt really help defcon at all, but its good to know that model loading is slowly
getting to a point where it can be used for 3D enviroments.