Renderer class now a true base class with virtual methods... I made the renderer class a base class that has virtual functions which can be overridden by the new OpenGL and DirectX11 renderers that inherit from Renderer.
Currently DirectX11 is a placeholders still but OpenGL functions just like before, there is no change.
Except the architecture is ready for dual renderers. The architecture was based off the idea that you
can just call a function from the renderer class, and whichever rendering context you are running the
rendering modules will override those virtual methods.

So calling SetBlendMode() from the renderer class will return either the OpenGL method or the DirectX11
method. The functionality actually somewhat matches how the window manager class works in some ways.

Some smaller changes are below:
- Prevent race condition when reinitialising the window, pressing apply in screen options window causing
  a crash as we are destroying the window while we are also trying to process an SDL event
- Added getters for the blend mode members
- Created empty d3d11 renderer class for when we start working on the implementation

Now the next step is most likely to actually work on creating the DirectX11 rendering layer, mapping all
OpenGL methods to DirectX11 equivalents. This will take a while and alot of pain and suffering. But it
will 100 percent be worth it in the end.