Added Rendering Abstraction to Resource and MegaVBO system... The bitmap, bitmapfont and image classes no longer directly call OpenGL functions in their methods. All of the OpenGL code that was being called from these classes have now been moved to the OpenGL renderer. Virtual functions were created in the base renderer class so that DirectX11 and OpenGL can use their own methods for texture creation.
Functionality is identical, OpenGL works just like before as all we have done is made the architecture
ready for multiple renderers that will need their own implementation for creating textures.

A list of smaller changes are below:
- Removed Blit() from the rendering system, call StaticSprite() or RotatingSprite()
  with immediate mode enabled in the function arguments from now on.
- Added some state caching to some OpenGL methods that were missing it before

Im not the biggest fan of having image creation inside the derived rendering classes, down the line
i might find a better solution to this but its really not the end of the world architecturally, as
we need the renderer to create these textures anyway.